+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/model/AppNotificationType.java
@@ -42,6 +42,10 @@ public class AppNotificationType extends HashMap<String, NotificationType> {
         put("ch.protonmail.android", NotificationType.GENERIC_EMAIL);
         put("security.pEp", NotificationType.GENERIC_EMAIL);
         put("eu.faircode.email", NotificationType.GENERIC_EMAIL);
+        put("net.thunderbird.android", NotificationType.GENERIC_EMAIL);
+        put("net.thunderbird.android.beta", NotificationType.GENERIC_EMAIL);
+        put("net.thunderbird.android.daily", NotificationType.GENERIC_EMAIL);
+        put("net.thunderbird.android.debug", NotificationType.GENERIC_EMAIL);
 
         // Generic SMS
         put("com.moez.QKSMS", NotificationType.GENERIC_SMS);